# app/internal/controller/search.go
package controller

import (
	"log"
	"net/http"

	"backend/app/internal/embedding"
	"backend/app/internal/models"

	"github.com/gin-gonic/gin"
)

type EmbeddingRequest = models.EmbeddingRequest

// Search godoc
//	@Summary		Search OpenAPI specifications
//	@Description	Retrieve OpenAPI specifications matching the given query
//	@Tags			search
//	@Accept			json
//	@Produce		json
//	@Param			fragment	body		models.EmbeddingRequest	true	"Search query"
//  TODO: Change this
//	@Success		200			{string}	OK
//	@Failure		400			{object}	models.HTTPError
//	@Router			/search [post]
func Search(ctx *gin.Context) {
	var body EmbeddingRequest
	err := ctx.BindJSON(&body)

	if err != nil {
		NewHTTPError(ctx, http.StatusBadRequest, "The query has not been correctly formatted")
	}

	embeddings := embedding.PerformPipeline([]string{body.Fragment}, true)
	log.Print(len(embeddings.Predictions[0]))
}